% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_domain_annotation.R
\name{annotate_domains}
\alias{annotate_domains}
\title{Annotate protein sequences with domains used for TF classification}
\usage{
annotate_domains(seq = NULL, evalue = 1e-05, threads = 1)
}
\arguments{
\item{seq}{An \code{AAStringSet} object. The sequences in this object must
represent only the translated sequences of primary (or longest) transcripts.
To create such an object from a FASTA file,
use \code{Biostrings::readAAStringSet()}.}

\item{evalue}{Numeric indicating the E-value threshold for \code{hmmsearch}
to be used for PlantTFDB domains without pre-defined domain cutoffs.
Default: 1e-05.}

\item{threads}{Numeric indicating the number of threads to be used
by \code{hmmsearch}. Default: 1.}
}
\value{
A list of 2 data frames named \strong{PlantTFDB} and
\strong{TAPscan}, with domain annotation for each classification scheme.
Both data frames have one row per domain hit and the following variables:
\describe{
\item{Gene}{Character, gene ID.}
\item{Domain}{Character, domain ID or name.}
\item{qlen}{Numeric, length of the profile HMM.}
\item{c_evalue}{Numeric, conditional E-value of the domain hit.}
\item{hmm_from, hmm_to}{Numeric, start and end coordinates of the alignment
in the profile HMM.}
}
Rows are kept in the same order as in the HMMER output, as this order
is used by the TAPscan classification algorithm.
}
\description{
Protein sequences are scanned with HMMER for domains used by the
classification schemes of \strong{PlantTFDB} and \strong{TAPscan}.
For PlantTFDB, PFAM and self-built profile HMMs are used with the
domain-specific score cutoffs defined by PlantTFDB (or an E-value threshold,
for domains without cutoffs).
For TAPscan, TAPscan v4 profile HMMs are used with their
gathering (GA) thresholds, as in the original TAPscan implementation.
}
\examples{
data(gsu)
seq <- gsu[1:5]
if(hmmer_is_installed()) {
    annotate_domains(seq)
}
}
