% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_annotation.R
\name{annotate_pfam}
\alias{annotate_pfam}
\title{Annotate proteins sequences with PFAM domains}
\usage{
annotate_pfam(seq = NULL, mode = "web")
}
\arguments{
\item{seq}{An AAStringSet object as returned
by \code{Biostrings::readAAStringSet()}.}

\item{mode}{Character specifying how to annotate PFAM domains. One of
'web' (default) or 'local'. If 'local', you must have hmmer installed in
your machine, and the command 'hmmsearch' must be in your PATH.}
}
\value{
A 2-column data frame with the variables \strong{Gene}
and \strong{Domain}, which contain gene IDs and domain IDs, respectively.
}
\description{
PFAM domains are assigned to each sequence using HMMER.
}
\examples{
data(gsu)
seq <- gsu[1:5]
annotate_pfam(seq)
}
