% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsu_annotation}
\alias{gsu_annotation}
\title{Domain annotation for the algae species Galdieria sulphuraria

The data set was created using the function \code{annotate_domains()}.}
\format{
A list of 2 data frames (\strong{PlantTFDB} and
\strong{TAPscan}) with domain annotation, as returned
by \code{annotate_domains()}.
}
\usage{
data(gsu_annotation)
}
\description{
Domain annotation for the algae species Galdieria sulphuraria

The data set was created using the function \code{annotate_domains()}.
}
\examples{
data(gsu_annotation)
}
\keyword{datasets}
