% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsu_families}
\alias{gsu_families}
\title{TFs families of the algae species Galdieria sulphuraria

The data set was created using the funcion \code{classify_tfs()}.}
\format{
A 2-column data frame with the following variables:
\describe{
\item{Gene}{Gene ID}
\item{Family}{TF family}
}
}
\usage{
data(gsu_families)
}
\description{
TFs families of the algae species Galdieria sulphuraria

The data set was created using the funcion \code{classify_tfs()}.
}
\examples{
data(gsu_families)
}
\keyword{datasets}
