% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planttfdb_scheme}
\alias{planttfdb_scheme}
\title{Data frame of PlantTFDB's TF family classification scheme}
\format{
A data frame with the following variables:
\describe{
\item{Family}{TF family name.}
\item{Subfamily}{TF subfamily name.}
\item{DBD}{DNA-binding domain}
\item{Auxiliary}{Auxiliary domain}
\item{Forbidden}{Forbidden domain}
}
}
\usage{
data(planttfdb_scheme)
}
\description{
The classification scheme is the same as the one used by PlantTFDB,
except for three family/subfamily names that were changed to match the
output of \code{classify_tfs()} ('M_type' to 'M-type', 'MYB_related' to
'MYB-related', and 'LBD (AS2/LOB)' to 'LBD'). As in PlantTFDB, families
without subfamilies have the family name as subfamily.
}
\examples{
data(planttfdb_scheme)
}
\references{
Jin, J., Tian, F., Yang, D. C., Meng, Y. Q., Kong, L., Luo, J., &
Gao, G. (2016). PlantTFDB 4.0: toward a central hub for transcription
factors and regulatory interactions in plants.
Nucleic acids research, gkw982.
}
\keyword{datasets}
