% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planttfhunter-package.R
\docType{package}
\name{planttfhunter-package}
\alias{planttfhunter}
\alias{planttfhunter-package}
\title{planttfhunter: Identification and classification of plant transcription factors}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

planttfhunter is used to identify plant transcription factors (TFs) from protein sequence data and classify them into families and subfamilies using the classification scheme implemented in PlantTFDB. TFs are identified using pre-built hidden Markov model profiles for DNA-binding domains. Then, auxiliary and forbidden domains are used with DNA-binding domains to classify TFs into families and subfamilies (when applicable). Currently, TFs can be classified in 58 different TF families/subfamilies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almeidasilvaf/planttfhunter}
  \item Report bugs at \url{https://support.bioconductor.org/t/planttfhunter}
}

}
\author{
\strong{Maintainer}: Fabrício Almeida-Silva \email{fabricio_almeidasilva@hotmail.com} (\href{https://orcid.org/0000-0002-5314-2964}{ORCID})

Authors:
\itemize{
  \item Yves Van de Peer \email{yves.vandepeer@psb.vib-ugent.be} (\href{https://orcid.org/0000-0003-4327-3730}{ORCID})
}

}
\keyword{internal}
