% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{score_cutoff}
\alias{score_cutoff}
\title{Data frame of domain cutoff for each domain}
\format{
A data frame with the following variables:
\describe{
\item{domain}{Character with the domain name.}
\item{domaincutoff}{Numeric with the domain score cutoff.}
}
}
\usage{
data(score_cutoff)
}
\description{
The cutoffs is the same as the one used by PlantTFDB. When domain cutoffs
are not specified (NA), the evalue cutoff in the \strong{evalue} parameter
of \code{annotate_pfam} is used.
}
\examples{
data(score_cutoff)
}
\references{
Jin, J., Tian, F., Yang, D. C., Meng, Y. Q., Kong, L., Luo, J., &
Gao, G. (2016). PlantTFDB 4.0: toward a central hub for transcription
factors and regulatory interactions in plants.
Nucleic acids research, gkw982.
}
\keyword{datasets}
