% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tap_class}
\alias{tap_class}
\title{Classes of transcription-associated proteins (TAPs) in TAPscan}
\format{
A data frame with the following variables:
\describe{
\item{TAP_family}{TAP family or subfamily, as in the variable
\strong{TAPscan_subfamily} of the output of \code{classify_tfs()}.}
\item{TAP_class}{TAP class. One of 'TF' (transcription factor),
'TR' (transcriptional regulator), or 'PT' (putative TAP).}
\item{Description}{Description of the TAP family.}
\item{References}{References for the TAP family, separated by ' | '.}
}
}
\usage{
data(tap_class)
}
\description{
Families of transcription-associated proteins (TAPs) in TAPscan v4, their
classes, descriptions, and references. Data were obtained from TAPscan's
website (file \emph{_data/import-tapinfo/tapinfo_v4.csv} in the
GitHub repository Rensing-Lab/TAPscan-v4-website). HTML tags in
descriptions and references were removed. Family MADS_MIKC, which is not
included in the original file, was manually added as a TF family (as
family MADS), without description and references.
}
\examples{
data(tap_class)
}
\references{
Petroll, R., Varshney, D., Hiltemann, S., Finke, H., Schreiber, M.,
de Vries, J., & Rensing, S. A. (2025). Enhanced sensitivity of TAPscan v4
enables comprehensive analysis of streptophyte transcription factor
evolution. The Plant Journal, 121(1), e17184.
}
\keyword{datasets}
