% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tapscan_scheme}
\alias{tapscan_scheme}
\title{Data frame of TAPscan's classification scheme}
\format{
A data frame with the following variables:
\describe{
\item{Family}{TAPscan family, as in variable \strong{TAPscan_family} of
the output of \code{classify_tfs()}.}
\item{Subfamily}{TAPscan subfamily, as in variable
\strong{TAPscan_subfamily} of the output of \code{classify_tfs()}.
Families without subfamilies have the family name as subfamily.}
\item{Rule}{Name of the rule in TAPscan's rules file.}
\item{Should}{Domains that genes should have, separated by ' and '.
Numbers in parentheses (e.g., 'Myb_DNA-binding (2)') indicate the
required number of copies of a domain.}
\item{Should_not}{Domains that genes should not have, separated
by ' or '.}
\item{TAP_class}{TAP class. One of 'TF' (transcription factor),
'TR' (transcriptional regulator), or 'PT' (putative TAP). NA for
families without a TAP class in TAPscan (see \code{tap_class}).}
}
}
\usage{
data(tapscan_scheme)
}
\description{
TAPscan v4 classification rules (v82), with one row per rule. Genes are
assigned to a family (and subfamily, if any) if they have all domains
in variable \strong{Should} and none of the domains in variable
\strong{Should_not}. For families with more than one rule (e.g., bZIP),
genes are assigned to the family if they satisfy any of the rules.
}
\examples{
data(tapscan_scheme)
}
\references{
Petroll, R., Varshney, D., Hiltemann, S., Finke, H., Schreiber, M.,
de Vries, J., & Rensing, S. A. (2025). Enhanced sensitivity of TAPscan v4
enables comprehensive analysis of streptophyte transcription factor
evolution. The Plant Journal, 121(1), e17184.
}
\keyword{datasets}
