% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_counts}
\alias{tf_counts}
\title{TF counts per family in 5 simulated species}
\format{
A SummarizedExperiment with TF frequencies per family in each
species in \strong{assay} and species metadata in \strong{colData}.
}
\usage{
data(tf_counts)
}
\description{
Simulated species were created by sampling 100 genes from the example
data set \strong{gsu} with after \code{set.seed(123)}.
}
\examples{
data(tf_counts)
}
\keyword{datasets}
