% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_tfs}
\alias{classify_tfs}
\title{Identify TFs and classify them in families}
\usage{
classify_tfs(domain_annotation = NULL)
}
\arguments{
\item{domain_annotation}{A 2-column data frame with the gene ID in the first
column and the domain ID in the second column.}
}
\value{
A 2-column data frame with the variables \strong{Gene}
and \strong{Family} representing gene ID and TF family, respectively.
}
\description{
Identify TFs and classify them in families
}
\examples{
data(gsu_annotation)
domain_annotation <- gsu_annotation
families <- classify_tfs(domain_annotation)
}
