% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsu}
\alias{gsu}
\title{Protein sequences of the algae species Galdieria sulphuraria}
\format{
An AAStringSet object as returned
by \code{Biostrings::readAAStringSet()}.
}
\usage{
data(gsu)
}
\description{
Data obtained from PLAZA Diatoms. Only genes containing domains used for
TF family classification were kept for package size issues.
}
\examples{
data(gsu)
}
\references{
Osuna-Cruz, C. M., Bilcke, G., Vancaester, E., De Decker, S., Bones, A. M.,
Winge, P., ... & Vandepoele, K. (2020). The Seminavis robusta genome
provides insights into the evolutionary adaptations of benthic diatoms.
Nature communications, 11(1), 1-13.
}
\keyword{datasets}
