% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsu_annotation}
\alias{gsu_annotation}
\title{Domain annotation for the algae species Galdieria sulphuraria

The data set was created using the funcion \code{annotate_pfam()} in local
mode.}
\format{
A 2-column data frame with the following variables:
\describe{
\item{Gene}{Gene ID}
\item{Annotation}{Domain ID or domain name when ID is not available in PFAM}
}
}
\usage{
data(gsu_annotation)
}
\description{
Domain annotation for the algae species Galdieria sulphuraria

The data set was created using the funcion \code{annotate_pfam()} in local
mode.
}
\examples{
data(gsu_annotation)
}
\keyword{datasets}
