% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hmmer_is_installed}
\alias{hmmer_is_installed}
\title{Check if HMMER is installed}
\usage{
hmmer_is_installed()
}
\value{
Logical indicating whether HMMER is installed or not.
}
\description{
Check if HMMER is installed
}
\examples{
hmmer_is_installed()
}
